change to lasttradedprice

+++ src/niftyswap-factory.ts
@@ -60,7 +60,7 @@ export function handleNewExchange(event: NewExchange): void {
     collection.nListedTokenIds = ZERO_BI;
     collection.nExchanges = ZERO_BI;
     collection.latestTradedTimestamp = ZERO_BI;
-    collection.latestTradedPrice = ZERO_BI.toBigDecimal()
+    collection.latestTradedPrice = ZERO_BI
   }
   collection.nExchanges = collection.nExchanges.plus(ONE_BI);
   collection.save();
@@ -84,7 +84,7 @@ export function handleNewExchange(event: NewExchange): void {
   niftyswapExchange.nSwaps = ZERO_BI;
   niftyswapExchange.nListedTokenIds = ZERO_BI;
   niftyswapExchange.latestTradedTimestamp = ZERO_BI
-  niftyswapExchange.latestTradedPrice = ZERO_BI.toBigDecimal()
+  niftyswapExchange.latestTradedPrice = ZERO_BI
   niftyswapExchange.save();
   currency.save();
